Rename fractal to aleph
