+++ gradle.properties
@@ -11,8 +11,8 @@ org.gradle.jvmargs=-Xmx1G
 
 # Mod Properties
 	mod_version = 1.0-SNAPSHOT
-	maven_group = com.fractalsmp
-	archives_base_name = fractal-carpet-addon
+	maven_group = net.alephsmp
+	archives_base_name = aleph-carpet
 
 # Dependencies
 	# currently not on the main fabric site, check on the maven: https://maven.fabricmc.net/net/fabricmc/fabric
