+++ src/main/java/net/alephsmp/aleph_carpet/AlephSimpleSettings.java
@@ -1,26 +1,26 @@
-package com.fractalsmp.fractal_carpet_addon;
+package net.alephsmp.aleph_carpet;
 
 import carpet.settings.Rule;
 import static carpet.settings.RuleCategory.FEATURE;
 
-public class FractalSimpleSettings {
-    public static final String FractalSettingCategory = "FractalAddon";
-    public static final String EndSettingsCategory = "FractalEndFeatures";
+public class AlephSimpleSettings {
+    public static final String AlephSettingsCategory = "aleph-addon";
+    public static final String EndSettingsCategory = "aleph-end-features";
     @Rule(
             desc = "Toggle for end gateway cooldown",
-            category = {FractalSettingCategory, FEATURE, EndSettingsCategory}
+            category = {AlephSettingsCategory, FEATURE, EndSettingsCategory}
     )
     public static boolean endGatewayCooldown = true;
 
     @Rule(
             desc = "Toggle for the main end island structure generation, turns off portal, egg, obsidian pillars, gateways and crystals",
-            category = {FractalSettingCategory, FEATURE, EndSettingsCategory}
+            category = {AlephSettingsCategory, FEATURE, EndSettingsCategory}
     )
     public static boolean endMainIslandStructureGen = true;
 
     @Rule(
             desc = "Toggle for end obsidian platform generation excluding players",
-            category = {FractalSettingCategory, FEATURE, EndSettingsCategory}
+            category = {AlephSettingsCategory, FEATURE, EndSettingsCategory}
     )
     public static boolean noObsidianPlatform = false;
 }
