+++ src/main/java/net/alephsmp/aleph_carpet/mixins/MinecraftClient_noopMixin.java
@@ -1,7 +1,7 @@
-package com.fractalsmp.fractal_carpet_addon.mixins;
+package net.alephsmp.aleph_carpet.mixins;
 
 
-import com.fractalsmp.fractal_carpet_addon.FractalExtension;
+import net.alephsmp.aleph_carpet.AlephExtension;
 import net.minecraft.client.MinecraftClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -12,6 +12,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class MinecraftClient_noopMixin {
     @Inject(method = "<init>", at = @At("RETURN"))
     private void loadExtension(CallbackInfo ci) {
-        FractalExtension.noop();
+        AlephExtension.noop();
     }
 }
