+++ src/main/java/net/alephsmp/aleph_carpet/mixins/MinecraftServer_noopMixin.java
@@ -1,6 +1,6 @@
-package com.fractalsmp.fractal_carpet_addon.mixins;
+package net.alephsmp.aleph_carpet.mixins;
 
-import com.fractalsmp.fractal_carpet_addon.FractalExtension;
+import net.alephsmp.aleph_carpet.AlephExtension;
 import net.minecraft.server.MinecraftServer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class MinecraftServer_noopMixin {
     @Inject(method = "<init>", at=@At("RETURN"))
     private void loadExtension(CallbackInfo ci) {
-        FractalExtension.noop();
+        AlephExtension.noop();
     }
 }
