+++ src/main/resources/assets/aleph_carpet/icon.png
