+++ src/main/resources/fabric.mod.json
@@ -1,20 +1,20 @@
 {
   "schemaVersion": 1,
-  "id": "fractal-carpet-addon",
+  "id": "aleph-carpet",
   "version": "1.0-SNAPSHOT",
 
-  "name": "Fabric Carpet Addon",
-  "description": "A carpet extension made for FractalSMP using MC 1.15.",
+  "name": "aleph-carpet",
+  "description": "A carpet extension made for AlephSMP in 1.16",
   "authors": [
     "FracturedCode"
   ],
   "contact": {
     "homepage": "https://TODO",
-    "sources": "https://github.com/FracturedCode/FractalCarpetAddon"
+    "sources": "https://github.com/FractalSMP/aleph-carpet"
   },
 
   "license": "MIT",
-  "icon": "assets/fractal_carpet_addon/icon.png",
+  "icon": "assets/aleph_carpet/icon.png",
 
   "environment": "*",
   "entrypoints": {
@@ -22,7 +22,7 @@
     ]
   },
   "mixins": [
-    "fractal_carpet_addon.mixins.json"
+    "aleph_carpet.mixins.json"
   ],
 
   "depends": {
